Group composite sinks
